+++ src/app/(board)/boards/_components/carousel/index.tsx
